bagitobjecttransfer/requirements.txt
alabaster==0.7.12
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
Babel==2.8.0
bagit==1.7.0
billiard==3.6.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Django==3.1.1
django-countries==6.1.3
django-formtools==2.2
django-rq==2.3.2
docutils==0.16
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.5.2
Jinja2==2.11.2
kombu==4.6.11
lazy-object-proxy==1.4.3
livereload==2.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
mysql-connector-python==8.0.22
packaging==20.4
pluggy==0.13.1
Pygments==2.7.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.24.0
rope==0.18.0
rq==1.5.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinx-autobuild==2020.9.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.10
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
